LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/values/shape-image-threshold-003-expected.txt


PASS shape-image-threshold can be assigned 'inherit' value
PASS shape-image-threshold is not inherited and defaults to 0
PASS shape-margin is not inherited
